requirements.txt
ace_tools==0.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
arelle-release==2.35.13
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
backoff==2.2.1
beautifulsoup4==4.12.3
benzinga==1.21
bleach==6.2.0
cattrs==24.1.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
comm==0.2.2
databento==0.47.0
databento-dbn==0.25.0
dataclasses-json==0.6.7
debugpy==1.8.9
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.7.0
earningscall==1.1.1
et_xmlfile==2.0.0
eventkit==1.0.3
-e git+https://github.com/faisalanjum/EventTrader.git@a256d138589f775f4bf22f015a00822598fef12a#egg=eventtrader
exchange_calendars==4.7
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.16.1
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.5.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.0
ib-insync==0.9.86
idna==3.10
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.30.0
ipywidgets==8.1.5
isodate==0.7.2
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.4
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
korean-lunar-calendar==0.3.1
lxml==5.3.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mistune==3.0.2
multidict==6.1.0
multitasking==0.0.11
mypy-extensions==1.0.0
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
neo4j==5.27.0
nest-asyncio==1.6.0
notebook==7.2.2
notebook_shim==0.2.4
numpy==2.1.3
openpyxl==3.1.5
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
peewee==3.17.8
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
polars==1.19.0
polygon==1.2.6
polygon-api-client==1.14.3
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.1
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.18.0
pyluach==2.2.0
pyparsing==3.2.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
redis==5.2.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-cache==1.2.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.22.0
schedule==1.2.2
Send2Trash==1.8.3
six==1.16.0
snakeviz==2.2.2
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
structlog==24.4.0
terminado==0.18.1
tinycss2==1.4.0
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241003
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
url-normalize==1.4.3
urllib3==2.3.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
widgetsnbextension==4.0.13
yarl==1.18.3
yfinance==0.2.51
zstandard==0.23.0
